requirements.txt
lxml==4.5.0
requests==2.22.0
selenium==3.141.0
python-dotenv==0.11.0
